requirements.txt
attrs==20.3.0
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
dukpy==0.2.3
echarts-countries-pypkg==0.1.6
folium==0.11.0
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
javascripthon==0.11
Jinja2==2.11.2
jupyter-echarts-pypkg==0.1.2
lml==0.0.2
macropy3==1.1.0b2
MarkupSafe==1.1.1
mock==4.0.2
numpy==1.19.4
packaging==20.4
pandas==1.1.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyahocorasick==1.4.0
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
six==1.15.0
toml==0.10.2
urllib3==1.25.11
zipp==3.4.0
